Allow use of tempfile for saving RDS files
With mngr_use_tempfile == TRUE, rather than directly writing RDS files out directly to disk, we first write the fil eout to a tempfile, and then move that file to the correct location. This helps when the ultimate location for the file is a slow (probably network) disk.

+++ R/files.R
@@ -76,9 +76,15 @@ save <- function(..., file){
 #' @param ... passed to default save
 #' @export
 saveRDS <- function(..., file){
-  unlink(file)
-  base:::saveRDS(..., file = file)
-
+  if (isTRUE(getOption("mngr_use_tempfile"))){
+    temp_file <- tempfile()
+    base:::saveRDS(..., file = temp_file)
+    file.rename(from = temp_file,
+                to = file)
+  } else {
+    unlink(file)
+    base:::saveRDS(..., file = file)
+}
 }
 
 #' pdf device
